refactor(xapian): Extract MailAddressInfo class and add tests
