+++ src/app/compose/draftdesk.service.spec.ts
@@ -19,7 +19,7 @@
 
 import {DraftFormModel} from './draftdesk.service';
 import { FromAddress } from '../rmmapi/from_address';
-import { MailAddressInfo } from '../xapian/messageinfo';
+import { MailAddressInfo } from '../common/mailaddressinfo';
 
 
 describe('DraftDesk', () => {
